Bestiary
Adds the bestiary, which adds every mob in the game including bosses as entries to a quest book chapter called bestiary which currently describes where they spawn, if they are from expert mode and in which dimension they're from. Fish are unlocked by bucketing, hostiles by killing, and animals by observing. All biomes are unlocked upon entering a dimension. I explain this in the colored instructions at the top.